changes datatype from int to sbyte for memory efficiency

+++ backend/backend/backend/src/Domain/Entities/InternShip.cs
@@ -11,7 +11,7 @@ public class InternShip
     public int Id { get; set; }
     public Unit Unit { get; set; }
     public string SchoolYear { get; set; }
-    public byte MaxStudents  { get; set; }
-    public int CurrentCountOfStudents { get; set; }
+    public sbyte MaxStudents  { get; set; }
+    public sbyte CurrentCountOfStudents { get; set; }
     public TrainingType RequiredTrainingType { get; set; }
 }
